changed so plot_markers can take in a bool series

+++ trtools/charting/charting.py
@@ -298,7 +298,7 @@ def process_signal(series, source):
     """
         Take any non 0/na value and changes it to corresponding value of source
     """
-    temp = series.copy()
+    temp = series.astype(float).copy()
     temp[temp == 0] = None
     temp *= source
     return temp
